Movie Detail Activity Data Binding
Addition of the Data Binding within the `MovieDetailActivity`.  Since this view currently is just flat xml, the only requirement is needed to perform data binding of the UI elements is to get the binding via the `getBinding` using the correct view (in this case the current focus provides the view).

No longer need to perform any `findById`.
Just get the binding and set the details w.r.t binding instance.