MovieDetailActivity Multiple AsyncTaskLoaders
Since database operations typically require the use of a Cursor and the `MovieDetailActivity` already had an implementation for the AsyncTaskLoader of type String, the AsyncTaskLoaders have been made into private classes within the `MovieDetailActivity`.  Originally was going to support an external but then will need to supply a public interface such that the activities can supply their own definitions of onFinished that the `onFinished` method within the loader would be responsible for actually calling.